+++ ui/src/pages/admin/settings/auth-providers/[id].tsx
@@ -1,9 +1,18 @@
 import React from "react";
-import { Form, Col, Row, Button, Alert, ButtonGroup } from "react-bootstrap";
+import {
+  Form,
+  Col,
+  Row,
+  Button,
+  Alert,
+  ButtonGroup,
+  InputGroup,
+} from "react-bootstrap";
 import {
   ChevronLeft as IconBack,
   Save as IconSave,
   Trash2 as IconDelete,
+  Edit2 as IconEdit,
 } from "react-feather";
 import { NextRouter } from "next/router";
 import FullLayout from "@/components/FullLayout";
@@ -15,12 +24,16 @@ import RuntimeConfig from "@/components/RuntimeConfig";
 import AuthProvider from "@/types/AuthProvider";
 import Ajax from "@/util/Ajax";
 import RedirectUtil from "@/util/RedirectUtil";
+import ErrorText from "@/types/ErrorText";
+import AjaxError from "@/util/AjaxError";
 
 interface State {
   loading: boolean;
   submitting: boolean;
   saved: boolean;
   goBack: boolean;
+  error: boolean;
+  errorText: string;
   name: string;
   providerType: number;
   authUrl: string;
@@ -33,6 +46,7 @@ interface State {
   userInfoLastnameField: string;
   clientId: string;
   clientSecret: string;
+  clientSecretEditing: boolean;
   logoutUrl: string;
   profilePageUrl: string;
   readOnly: boolean;
@@ -53,6 +67,8 @@ class EditAuthProvider extends React.Component<Props, State> {
       submitting: false,
       saved: false,
       goBack: false,
+      error: false,
+      errorText: "",
       name: "",
       providerType: 0,
       authUrl: "",
@@ -65,6 +81,7 @@ class EditAuthProvider extends React.Component<Props, State> {
       userInfoLastnameField: "",
       clientId: "",
       clientSecret: "",
+      clientSecretEditing: true,
       logoutUrl: "",
       profilePageUrl: "",
       readOnly: false,
@@ -79,10 +96,11 @@ class EditAuthProvider extends React.Component<Props, State> {
     this.loadData();
   };
 
-  loadData = () => {
+  loadData = async () => {
     const { id } = this.props.router.query;
     if (id && typeof id === "string" && id !== "add") {
-      AuthProvider.get(id).then((authProvider) => {
+      try {
+        const authProvider = await AuthProvider.get(id);
         this.entity = authProvider;
         this.setState({
           name: authProvider.name,
@@ -97,12 +115,15 @@ class EditAuthProvider extends React.Component<Props, State> {
           userInfoLastnameField: authProvider.userInfoLastnameField,
           clientId: authProvider.clientId,
           clientSecret: authProvider.clientSecret,
+          clientSecretEditing: false,
           logoutUrl: authProvider.logoutUrl,
           profilePageUrl: authProvider.profilePageUrl,
           readOnly: authProvider.readOnly,
           loading: false,
         });
-      });
+      } catch {
+        this.setState({ loading: false, error: true });
+      }
     } else {
       this.setState({
         loading: false,
@@ -110,7 +131,7 @@ class EditAuthProvider extends React.Component<Props, State> {
     }
   };
 
-  onSubmit = (e: any) => {
+  onSubmit = async (e: any) => {
     e.preventDefault();
     this.entity.name = this.state.name;
     this.entity.providerType = this.state.providerType;
@@ -123,24 +144,42 @@ class EditAuthProvider extends React.Component<Props, State> {
     this.entity.userInfoFirstnameField = this.state.userInfoFirstnameField;
     this.entity.userInfoLastnameField = this.state.userInfoLastnameField;
     this.entity.clientId = this.state.clientId;
-    this.entity.clientSecret = this.state.clientSecret;
+    if (this.state.clientSecretEditing) {
+      this.entity.clientSecret = this.state.clientSecret;
+    }
     this.entity.logoutUrl = this.state.logoutUrl;
     this.entity.profilePageUrl = this.state.profilePageUrl;
-    this.entity.save().then(() => {
+
+    try {
+      await this.entity.save();
       this.props.router.push(
         "/admin/settings/auth-providers/" + this.entity.id,
       );
       this.setState({
         saved: true,
+        submitting: false,
+        clientSecretEditing: false,
       });
-    });
+    } catch (e) {
+      let code: number = 0;
+      if (e instanceof AjaxError) {
+        code = e.appErrorCode;
+      }
+      this.setState({
+        error: true,
+        errorText: code ? ErrorText.getTextForAppCode(code, this.props.t) : "",
+      });
+    }
   };
 
-  deleteItem = () => {
+  deleteItem = async () => {
     if (window.confirm(this.props.t("confirmDeleteAuthProvider"))) {
-      this.entity.delete().then(() => {
+      try {
+        await this.entity.delete();
         this.setState({ goBack: true });
-      });
+      } catch {
+        this.setState({ error: true });
+      }
     }
   };
 
@@ -242,9 +281,15 @@ class EditAuthProvider extends React.Component<Props, State> {
     let hint = <></>;
     if (this.state.saved) {
       hint = <Alert variant="success">{this.props.t("entryUpdated")}</Alert>;
+    } else if (this.state.error) {
+      hint = (
+        <Alert variant="danger">
+          {this.state.errorText ?? this.props.t("errorSave")}
+        </Alert>
+      );
     }
 
-    let urlInfo = <></>;
+    let callbackUrlInfo = <></>;
     let buttonDelete = (
       <Button
         className="btn-sm"
@@ -272,7 +317,7 @@ class EditAuthProvider extends React.Component<Props, State> {
           {backButton} {buttonDelete} {buttonSave}
         </>
       );
-      urlInfo = (
+      callbackUrlInfo = (
         <Form.Group as={Row}>
           <Form.Label column sm="2">
             Callback URL
@@ -302,6 +347,36 @@ class EditAuthProvider extends React.Component<Props, State> {
       <FullLayout headline={this.props.t("editAuthProvider")} buttons={buttons}>
         <Form onSubmit={this.onSubmit} id="form">
           {hint}
+          <Form.Group as={Row} hidden={this.entity.id !== ""}>
+            <Form.Label column sm="2">
+              {this.props.t("templates")}
+            </Form.Label>
+            <Col sm="9">
+              <ButtonGroup>
+                <Button
+                  variant="outline-secondary"
+                  onClick={this.templateGoogle}
+                >
+                  Google
+                </Button>
+                <Button
+                  variant="outline-secondary"
+                  onClick={this.templateMicrosoft}
+                >
+                  Microsoft
+                </Button>
+                <Button
+                  variant="outline-secondary"
+                  onClick={this.templateKeycloak}
+                >
+                  Keycloak
+                </Button>
+                <Button variant="outline-secondary" onClick={this.templateOkta}>
+                  Okta
+                </Button>
+              </ButtonGroup>
+            </Col>
+          </Form.Group>
           <Form.Group as={Row}>
             <Form.Label column sm="2">
               {this.props.t("name")}
@@ -420,17 +495,40 @@ class EditAuthProvider extends React.Component<Props, State> {
               Client Secret
             </Form.Label>
             <Col sm="9">
-              <Form.Control
-                type="text"
-                placeholder="Client Secret"
-                value={this.state.clientSecret}
-                onChange={(e: any) =>
-                  this.setState({ clientSecret: e.target.value })
-                }
-                required={true}
-                pattern="[^\s]+"
-                title="Client Secret (whitespaces are not allowed)"
-              />
+              {!this.state.clientSecretEditing && this.entity.id ? (
+                <InputGroup>
+                  <Form.Control
+                    type="text"
+                    value="••••••••••••••••"
+                    readOnly={true}
+                  />
+                  <Button
+                    variant="outline-secondary"
+                    onClick={() =>
+                      this.setState({
+                        clientSecretEditing: true,
+                        clientSecret: "",
+                      })
+                    }
+                    title={this.props.t("edit")}
+                  >
+                    <IconEdit className="feather" />
+                  </Button>
+                </InputGroup>
+              ) : (
+                <Form.Control
+                  type="text"
+                  placeholder="Client Secret"
+                  value={this.state.clientSecret}
+                  onChange={(e: any) =>
+                    this.setState({ clientSecret: e.target.value })
+                  }
+                  required={true}
+                  pattern="[^\s]+"
+                  title="Client Secret (whitespaces are not allowed)"
+                  autoFocus={this.state.clientSecretEditing && !!this.entity.id}
+                />
+              )}
             </Col>
           </Form.Group>
           <Form.Group as={Row}>
@@ -525,37 +623,7 @@ class EditAuthProvider extends React.Component<Props, State> {
               />
             </Col>
           </Form.Group>
-          {urlInfo}
-          <Form.Group as={Row} hidden={this.entity.id !== ""}>
-            <Form.Label column sm="2">
-              {this.props.t("templates")}
-            </Form.Label>
-            <Col sm="9">
-              <ButtonGroup>
-                <Button
-                  variant="outline-secondary"
-                  onClick={this.templateGoogle}
-                >
-                  Google
-                </Button>
-                <Button
-                  variant="outline-secondary"
-                  onClick={this.templateMicrosoft}
-                >
-                  Microsoft
-                </Button>
-                <Button
-                  variant="outline-secondary"
-                  onClick={this.templateKeycloak}
-                >
-                  Keycloak
-                </Button>
-                <Button variant="outline-secondary" onClick={this.templateOkta}>
-                  Okta
-                </Button>
-              </ButtonGroup>
-            </Col>
-          </Form.Group>
+          {callbackUrlInfo}
         </Form>
       </FullLayout>
     );